rag2/requirements.txt
sentence-transformers
Wikipedia-API
python-dotenv
openai
numpy